+++ app/src/pages/Dashboard/EventsOverviewPage/EventsOverviewPage.tsx
@@ -3,23 +3,16 @@
 // SPDX-License-Identifier: EUPL-1.2
 import { IconButton, Skeleton, Stack, styled, Typography } from '@mui/material';
 import { ArrowDownIcon, formatDate } from '@opentalk/common';
-import {
-  CursorPaginated,
-  DateTime,
-  Event,
-  EventException,
-  InviteStatus,
-  isEvent,
-  isTimelessEvent,
-} from '@opentalk/rest-api-rtk-query';
-import { createSelector } from '@reduxjs/toolkit';
-import { endOfISOWeek, getWeek, startOfISOWeek } from 'date-fns';
+import { DateTime, Event, EventException, InviteStatus, isTimelessEvent } from '@opentalk/rest-api-rtk-query';
+import { endOfISOWeek, getWeek, isAfter, isBefore, isEqual, startOfISOWeek } from 'date-fns';
 import i18n, { t } from 'i18next';
-import { partition, groupBy } from 'lodash';
+import { groupBy } from 'lodash';
 import React, { useEffect, useMemo, useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useGetEventsQuery } from '../../../api/rest';
+import { Toggle } from '../../../commonComponents';
+import { selectFilteredEventByFavorite } from '../../../store/selectors';
 import { useHeader } from '../../../templates/DashboardTemplate';
 import { getExpandedEvents } from '../../../utils/eventUtils';
 import EventsOverview from './fragments/EventsOverview';
@@ -79,15 +72,45 @@ export const filterByTimePeriod = (timePeriod: DashboardEventsFilters['timePerio
   }
 };
 
+enum TimePerspectiveFilter {
+  TimeIndependent = 'timeindependent',
+  Future = 'future',
+  Past = 'past',
+}
+
+interface ToggleOptionsProps {
+  value: TimePerspectiveFilter;
+  label: string;
+}
+
 const EventsOverviewPage = ({ header }: MeetingsPageProps) => {
-  const [expandAll, setExpandAll] = useState<boolean>(true);
+  const [expandAccordion, setExpandAccordion] = useState<string>('');
   const [filter, setFilter] = useState<DashboardEventsFilters>({
     timePeriod: TimeFilter.Month,
     timeMin: new Date().toTimeString() as DateTime,
     openInvitedMeeting: false,
     favoriteMeetings: false,
   });
+
+  const [timePeriodFilter, setTimePeriodFilter] = useState<TimePerspectiveFilter>(
+    TimePerspectiveFilter.TimeIndependent
+  );
+
   const { t } = useTranslation();
+  const toggleOptions: Array<ToggleOptionsProps> = [
+    {
+      value: TimePerspectiveFilter.TimeIndependent,
+      label: t('dashboard-meeting-details-page-time-independent'),
+    },
+    {
+      value: TimePerspectiveFilter.Future,
+      label: t('dashboard-meeting-details-page-future'),
+    },
+    {
+      value: TimePerspectiveFilter.Past,
+      label: t('dashboard-meeting-details-page-past'),
+    },
+  ];
 
   const { setHeader } = useHeader();
 
@@ -105,50 +128,75 @@ const EventsOverviewPage = ({ header }: MeetingsPageProps) => {
     };
   }, [header, setHeader, filter]);
 
-  const selectFilteredEvents = useMemo(() => {
-    return createSelector(
-      [(state: { data: CursorPaginated<EventException | Event> | undefined }) => state.data],
-      (data): MeetingsProp[] | undefined => {
-        if (data === undefined) {
-          return data;
-        }
-        const { data: selectedEvents } = data;
-        const expandedEvents = getExpandedEvents(selectedEvents, true);
-
-        const invitedMeetings = expandedEvents?.filter((event) =>
-          isEvent(event) ? event.inviteStatus !== InviteStatus.Accepted : false
-        );
-        const inviteStatusFilteredData = filter.openInvitedMeeting ? invitedMeetings : expandedEvents;
-        const partitionList = partition(inviteStatusFilteredData, (event) => event.isTimeIndependent);
-
-        const timeIndependent =
-          partitionList[0].length > 0
-            ? [{ title: t('dashboard-event-time-independent-meetings'), events: [...partitionList[0]] }]
-            : [];
-
-        const timeDependent = groupBy(partitionList[1], (event) =>
-          filterByTimePeriod(
-            filter.timePeriod,
-            isTimelessEvent(event) ? event.createdAt : (event.startsAt?.datetime as DateTime)
-          )
-        );
-
-        const timeDependentArray = Object.entries(timeDependent).map(([title, events]) => ({ title, events }));
-
-        return [...timeIndependent, ...timeDependentArray];
-      }
+  const formatItemsAsMeetingProps = (event: Array<Event>) => {
+    const groupByTimeFilter = groupBy(event, (event) =>
+      filterByTimePeriod(
+        filter.timePeriod,
+        isTimelessEvent(event) ? event.createdAt : (event.startsAt?.datetime as DateTime)
+      )
     );
-  }, [t, filter.timePeriod, filter.openInvitedMeeting]);
+    const groupByTimeFilterArray = Object.entries(groupByTimeFilter).map(([title, events]) => ({ title, events }));
+    return [...groupByTimeFilterArray];
+  };
+
+  const formatedEventsByHeaderChange = useMemo(() => {
+    {
+      return (events: Array<Event>) => {
+        if (filter.favoriteMeetings) {
+          return events.filter((events) => events.isFavorite);
+        }
+        if (filter.openInvitedMeeting) {
+          return events.filter((events) => events.inviteStatus !== InviteStatus.Accepted);
+        }
+        return events;
+      };
+    }
+  }, [filter.favoriteMeetings, filter.openInvitedMeeting]);
 
   const { filteredData, isLoading } = useGetEventsQuery(
     { favorites: filter.favoriteMeetings, perPage: 100, adhoc: false },
     {
       selectFromResult: ({ data, isLoading }) => ({
-        filteredData: selectFilteredEvents({ data }),
+        filteredData: selectFilteredEventByFavorite({ data }),
         isLoading,
       }),
     }
   );
+
+  const evenData = useMemo(() => {
+    const expandedEvents = formatedEventsByHeaderChange(getExpandedEvents(filteredData, true));
+    let currentFilteredEvents: Array<Event> = [];
+
+    if (timePeriodFilter === TimePerspectiveFilter.TimeIndependent) {
+      currentFilteredEvents = expandedEvents.filter((event) => isTimelessEvent(event));
+      const constructMeetingProp = {
+        title: t('dashboard-meeting-details-page-time-independent'),
+        events: currentFilteredEvents,
+      };
+      setExpandAccordion('all');
+      return [constructMeetingProp];
+    } else if (timePeriodFilter === TimePerspectiveFilter.Future) {
+      const currentDate = new Date();
+      currentFilteredEvents = expandedEvents.filter((event) => {
+        return (
+          (!isTimelessEvent(event) && isEqual(new Date(event.startsAt.datetime), currentDate)) ||
+          (!isTimelessEvent(event) && isAfter(new Date(event.startsAt.datetime), currentDate))
+        );
+      });
+      const events = formatItemsAsMeetingProps(currentFilteredEvents);
+      if (events.length > 0) {
+        setExpandAccordion(events[0].title);
+      }
+      return events;
+    } else {
+      const currentDate = new Date();
+      currentFilteredEvents = expandedEvents.filter(
+        (event) => !isTimelessEvent(event) && isBefore(new Date(event.startsAt.datetime), currentDate)
+      );
+      setExpandAccordion('');
+      return formatItemsAsMeetingProps(currentFilteredEvents);
+    }
+  }, [t, filter, timePeriodFilter, filteredData]);
 
   if (isLoading) {
     return (
@@ -166,11 +214,20 @@ const EventsOverviewPage = ({ header }: MeetingsPageProps) => {
         <Typography variant={'h1'} component={'h2'}>
           {t('dashboard-events-my-meetings')}
         </Typography>
-        <ArrowDownButton active={expandAll} onClick={() => setExpandAll((prev) => !prev)}>
+        <Toggle
+          options={toggleOptions}
+          onChange={(timePeriod: TimePerspectiveFilter) => {
+            setTimePeriodFilter(timePeriod);
+          }}
+        ></Toggle>
+        <ArrowDownButton
+          active={expandAccordion === 'all'}
+          onClick={() => setExpandAccordion((prev) => (prev === 'all' ? '' : 'all'))}
+        >
           <ArrowDownIcon color="secondary" />
         </ArrowDownButton>
       </Stack>
-      <EventsOverview data-testid="EventsOverview" entries={filteredData || []} expandAll={expandAll} />
+      <EventsOverview data-testid="EventsOverview" entries={evenData || []} expandAccordion={expandAccordion} />
     </Stack>
   );
 };